% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.n2.R
\docType{package}
\name{heatmapGen2}
\alias{heatmapGen2}
\alias{heatmapGen2-package}
\title{Heatmap Tools}

\details{
\tabular{ll}{
Package: \tab heatmapGen2\cr
Type: \tab Package\cr
Initial version: \tab 0.1-0\cr
Created: \tab 2016-08-26\cr
License: \tab MIT License\cr
}
The key function in the package for end users is \code{heatmap.n2} and there are several internal functions.
}
\author{
Vidal Fey <vidal.fey@utu.fi> based on code created by Henri Sara and maintained by Dhanaprakash Jambulingam <dhajam@utu.fi>
}
\keyword{package}

