\name{hmcols}
\alias{hmcols}

\title{
  Get a green-black-red heatmap palette
}
\description{
  A function that returns a green-black-red palette for use in heatmaps etc.
}
\usage{
hmcols()
}
\value{
  Vector of color strings.
}
\note{
The result should be cached in a variable for efficiency if used often.
This should perhaps be a data object rather than a function.
}
\keyword{ color }
